+++ frontend/src/container/BillingContainer/BillingContainer.tsx
@@ -20,7 +20,7 @@ import { ErrorResponse, SuccessResponse } from 'types/api';
 import { CheckoutSuccessPayloadProps } from 'types/api/billing/checkout';
 import { License } from 'types/api/licenses/def';
 import AppReducer from 'types/reducer/app';
-import { getFormattedDate } from 'utils/timeUtils';
+import { getFormattedDate, getRemainingDays } from 'utils/timeUtils';
 
 interface DataType {
 	key: string;
@@ -97,19 +97,6 @@ const dummyColumns: ColumnsType<DataType> = [
 		render: renderSkeletonInput,
 	},
 ];
-
-export const getRemainingDays = (billingEndDate: number): number => {
-	// Convert Epoch timestamps to Date objects
-	const startDate = new Date(); // Convert seconds to milliseconds
-	const endDate = new Date(billingEndDate * 1000); // Convert seconds to milliseconds
-
-	// Calculate the time difference in milliseconds
-	// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-	// @ts-ignore
-	const timeDifference = endDate - startDate;
-
-	return Math.ceil(timeDifference / (1000 * 60 * 60 * 24));
-};
 
 export default function BillingContainer(): JSX.Element {
 	const daysRemainingStr = 'days remaining in your billing period.';
